This patch introduces the PyObjCTools package, and moves Foundation.Conversion and AppKit.NibClassBuilder to that package. It also adds a new module containing the runEvenloop function proposed by Just (includeing the improvements proposed by Jack, and some bugs introduced by me).
The examples and project-templates have been adapted to this change.

Also includes a fix for 'NSMachPort.alloc()', which used to crash: We no longer
call 'retain' on the result of 'alloc' messages. Because of this I've also
enabled the option that updates the retainCount for methods that return
non-autoreleased objects (e.g. copy and friends). This should keep live
simpler.


git-svn-id: http://svn.red-bean.com/pyobjc/trunk@499 f55f28a5-9edb-0310-a011-a803cfcd5d25