Updated protocol handling. It is no longer necessary to specify the list of protocols defined by a class, the bridge will calculate this list based on the methods implemented by a class.
Also fix the definition of a number of protocols.

This implements some more methods on NS{Mutable,}Dictionary to make it more
dict-like. The 'poseAsClass:' method is disabled because it crashes the brige due
to the implementation of that method in Objective-C. Given the other problems of
poseAs I don't think it would be usefull to fix this right now.


git-svn-id: http://svn.red-bean.com/pyobjc/trunk@605 f55f28a5-9edb-0310-a011-a803cfcd5d25