This fixes the symptoms(sp?) of a bug, not sure if this is the correct fix...
The bug: I sometimes get 'SystemError: error return without exception set' when
I have a typo in the name of an initialiser method. This never happens on the
toplevel, but in a callback. The bug is always triggered by code like this:

	foo = SomeClass.alloc().inti() # Oops can't spell init

The "fix" saves the exception information before actually deallocating the
uninitialized Objective-C object and restores that information afterwards.


git-svn-id: http://svn.red-bean.com/pyobjc/trunk@819 f55f28a5-9edb-0310-a011-a803cfcd5d25